backend/requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
pdfplumber==0.11.9
openai>=1.0.0
python-multipart==0.0.22
nltk
python-dotenv
